refactor code, add widgets.ts
